+++ examples/eth-pm-wallet-encryption/src/crypto.ts
@@ -1,7 +1,7 @@
 import "@ethersproject/shims";
 
 import { PublicKeyMessage } from "./messaging/wire";
-import { hexToBytes, equalByteArrays, bufToHex } from "js-waku/lib/utils";
+import { hexToBytes, equalByteArrays, bytesToHex } from "js-waku/lib/utils";
 import * as sigUtil from "eth-sig-util";
 
 /**
@@ -41,7 +41,7 @@ function buildMsgParams(encryptionPublicKey: Uint8Array, fromAddress: string) {
       version: "1",
     },
     message: {
-      encryptionPublicKey: bufToHex(encryptionPublicKey),
+      encryptionPublicKey: bytesToHex(encryptionPublicKey),
       ownerAddress: fromAddress,
     },
     // Refers to the keys of the *types* object below.
@@ -88,13 +88,13 @@ export async function signEncryptionKey(
 export function validatePublicKeyMessage(msg: PublicKeyMessage): boolean {
   const recovered = sigUtil.recoverTypedSignature_v4({
     data: JSON.parse(
-      buildMsgParams(msg.encryptionPublicKey, "0x" + bufToHex(msg.ethAddress))
+      buildMsgParams(msg.encryptionPublicKey, "0x" + bytesToHex(msg.ethAddress))
     ),
-    sig: "0x" + bufToHex(msg.signature),
+    sig: "0x" + bytesToHex(msg.signature),
   });
 
   console.log("Recovered", recovered);
-  console.log("ethAddress", "0x" + bufToHex(msg.ethAddress));
+  console.log("ethAddress", "0x" + bytesToHex(msg.ethAddress));
 
   return equalByteArrays(recovered, msg.ethAddress);
 }